Split metadata generator out from file operation functions
