+++ src/core/task-persistence/index.ts
@@ -1,2 +1,3 @@
 export { readApiMessages, saveApiMessages } from "./apiMessages"
 export { readTaskMessages, saveTaskMessages } from "./taskMessages"
+export { taskMetadata } from "./taskMetadata"
